Place processing-related code into separate file In order for messages to be processed before the chat screen is open, all messages should be processed in their own file, potentially on a separate thread. This may improve performance for difficult to process messages, and should make the system more modular and easier to modify in the future (such as changing the chat screen).
Place processing-related code into separate file.
