+++ Chat/EventArgsData.cs
@@ -7,12 +7,18 @@
     public class MessageReceivedEventArgs : EventArgs
     {
         public Client client;
-        public Message message;
+        public uint messageId;
+        public Message.MessageTypes messageType;
+        public string messageText;
+        public byte[] messageBytes;
 
-        public MessageReceivedEventArgs(Client client, Message message)
+        public MessageReceivedEventArgs(Client client, uint messageId, Message.MessageTypes messageType, string messageText, byte[] messageBytes)
         {
             this.client = client;
-            this.message = message;
+            this.messageId = messageId;
+            this.messageType = messageType;
+            this.messageText = messageText;
+            this.messageBytes = messageBytes;
         }
     }
