+++ Chat/frmChatScreen.cs
@@ -6,25 +6,21 @@ namespace Chat
     {
         private bool askToClose = true;
 
-        private delegate void MessageDelegate(object sender, MessageReceivedEventArgs e);
         private delegate void HeartbeatDelegate(Client client);
         private delegate void ClearClientClistDelegate();
         private delegate void AddClientToClientListDelegate(string username);
         private delegate void PrintChatMessageDelegate(string text);
-        private delegate void NextConnectionSetupStepDelegate(Client client);
         private delegate DialogResult ShowMessageBoxDelegate(string message, string caption, MessageBoxButtons messageBoxButtons, MessageBoxIcon messageBoxIcon);
 
-        private string kickFormat = "/kick [Username] [Reason (optional)]";
-        private string adminFormat = "/admin [Username] [True/False (optional)]";
+        private delegate void OpenMainMenuDelegate();
 
         public FrmChatScreen()
         {
             InitializeComponent();
-            FrmHolder.network = new Network();
-            SetNetworkEventHandlers();
+            FrmHolder.processing = new Processing();
+            SetProcessingEventHandlers();
             SetFormEventHandlers();
             xlsvConnectedUsers.Columns[0].Width = xlsvConnectedUsers.Width - 5;
-            FrmHolder.network.BeginNetworkThreads();
         }
 
         private void SetFormEventHandlers()
@@ -33,352 +29,14 @@ namespace Chat
             this.FormClosing += new FormClosingEventHandler(OnClosing);
         }
 
-        private void SetNetworkEventHandlers()
+        private void SetProcessingEventHandlers()
         {
-            FrmHolder.network.MessageReceivedEvent += OnMessageReceived;
-            FrmHolder.network.HeartbeatTimeoutEvent += OnHeartbeatTimeoutFailure;
-            FrmHolder.network.PrintChatMessageEvent += OnPrintChatMessage;
-            FrmHolder.network.ClearClientListEvent += OnClearClientList;
-            FrmHolder.network.AddClientToClientListEvent += OnAddClientToClientList;
-            FrmHolder.network.NextConnectionSetupStep += OnNextConnectionSetupStep;
-            FrmHolder.network.ShowMessageBoxEvent += OnShowMessagBox;
-        }
-
-        private void ProcessMessage(object sender, MessageReceivedEventArgs e)
-        {
-            e.client.heartbeatReceieved = true;
-            e.client.heartbeatFailures = 0;
-            if (e.message.messageBytes != null)
-            {
-                if (e.message.CheckIfCanConvertToText())
-                {
-                    e.message.MessageTextToOrFromBytes();
-                }
-            }
-#if DEBUG && messageReceivedUpdates
-            if (e.message.messageType != Message.MessageTypes.Heartbeat)
-            {
-                if (e.message.messageText != null)
-                {
-                    PrintChatMessage($"[RECEIVED] Type: {e.message.messageType}. ID: {e.message.messageId}. Text: {e.message.messageText}");
-                }
-                else
-                {
-                    PrintChatMessage($"[RECEIVED] Type: {e.message.messageType}. ID: {e.message.messageId}");
-                }
-            }
-#endif
-            if (e.message.messageType != Message.MessageTypes.Acknowledgement && e.message.messageType != Message.MessageTypes.ClientDisconnect && e.message.messageType != Message.MessageTypes.Heartbeat)
-            {
-                FrmHolder.network.BeginWrite(e.client, FrmHolder.network.ComposeMessage(e.client, e.message.messageId, Message.MessageTypes.Acknowledgement, null, null)); // Acknowledge received message
-                if (e.client.connectionSetupComplete)
-                {
-                    foreach (Message alreadyReceivedMessage in e.client.messagesReceived)
-                    {
-                        if (e.message.messageId == alreadyReceivedMessage.messageId)
-                        {
-                            return;
-                        }
-                    }
-                }
-                e.client.messagesReceived.Add(e.message);
-            }
-
-            if (e.message.messageType == Message.MessageTypes.None)
-            {
-                return;
-            }
-            else if (e.message.messageType == Message.MessageTypes.Acknowledgement)
-            {
-                foreach (Message item in e.client.messagesToBeSent)
-                {
-                    if (item.messageId == e.message.messageId)
-                    {
-                        e.client.messagesSent.Add(item);
-                        e.client.messagesToBeSent.Remove(item);
-                        break;
-                    }
-                }
-                if (e.client.sendingMessageQueue)
-                {
-                    FrmHolder.network.SendFirstMessageInMessageQueue(e.client);
-                }
-            }
-            else if (e.message.messageType == Message.MessageTypes.ChatMessage)
-            {
-                string[] parts = e.message.messageText.Split(' ', 2);
-                string username = parts[0];
-                string messageText = parts[1];
-                PrintChatMessage($"{username}: {messageText}");
-                if (FrmHolder.hosting)
-                {
-                    FrmHolder.network.SendToAll(null, Message.MessageTypes.ChatMessage, e.message.messageText, null);
-                }
-            }
-            else if (e.message.messageType == Message.MessageTypes.ClientDisconnect)
-            {
-                if (FrmHolder.hosting)
-                {
-                    FrmHolder.network.connectedClients.Remove(e.client);
-                    if (e.client.tcpClient != null)
-                    {
-                        e.client.tcpClient.Close();
-                    }
-                    if (e.client.username != null)
-                    {
-                        PrintChatMessage($"{e.client.username} disconnected");
-                        FrmHolder.network.SendToAll(null, Message.MessageTypes.UserDisconnected, e.client.username, null);
-                    }
-                    FrmHolder.network.UpdateClientLists();
-                }
-                else
-                {
-                    if (FrmHolder.network.clientThread != null && FrmHolder.network.clientThread.IsAlive)
-                    {
-                        FrmHolder.network.clientCancellationTokenSource.Cancel();
-                    }
-                    MessageBox.Show("The server was closed.", "Information", MessageBoxButtons.OK, MessageBoxIcon.Information);
-                    OpenMainMenu();
-                }
-            }
-            else if (e.message.messageType == Message.MessageTypes.UsernameInUse)
-            {
-                FrmHolder.network.clientCancellationTokenSource.Cancel();
-                MessageBox.Show("This username is already in use", "Information", MessageBoxButtons.OK, MessageBoxIcon.Information);
-                OpenMainMenu();
-            }
-            else if (e.message.messageType == Message.MessageTypes.UserConnected)
-            {
-                PrintChatMessage($"{e.message.messageText} connected");
-            }
-            else if (e.message.messageType == Message.MessageTypes.UserDisconnected)
-            {
-                PrintChatMessage($"{e.message.messageText} disconnected");
-            }
-            else if (e.message.messageType == Message.MessageTypes.ClearUserList)
-            {
-                xlsvConnectedUsers.Items.Clear();
-            }
-            else if (e.message.messageType == Message.MessageTypes.AddToUserList)
-            {
-                xlsvConnectedUsers.Items.Add(e.message.messageText);
-            }
-            else if (e.message.messageType == Message.MessageTypes.Kicked)
-            {
-                if (FrmHolder.network.clientThread != null && FrmHolder.network.clientThread.IsAlive)
-                {
-                    FrmHolder.network.clientCancellationTokenSource.Cancel();
-                }
-                string[] parts = e.message.messageText.Split(' ', 2);
-                string username = parts[0];
-                string reason = parts[1];
-                reason = reason.Trim();
-                if (!string.IsNullOrWhiteSpace(reason))
-                {
-                    MessageBox.Show($"You have been kicked by {username} with reason: {reason}", "Information", MessageBoxButtons.OK, MessageBoxIcon.Information);
-                }
-                else
-                {
-                    MessageBox.Show($"You have been kicked by {username}", "Information", MessageBoxButtons.OK, MessageBoxIcon.Information);
-                }
-                OpenMainMenu();
-            }
-            else if (e.message.messageType == Message.MessageTypes.OtherUserKicked)
-            {
-                string[] parts = e.message.messageText.Split(' ', 3);
-                string username = parts[0];
-                string kickerUsername = parts[1];
-                string reason = parts[2];
-                reason = reason.Trim();
-                if (!string.IsNullOrWhiteSpace(reason))
-                {
-                    PrintChatMessage($"{username} was kicked by {kickerUsername} with reason: {reason}");
-                }
-                else
-                {
-                    PrintChatMessage($"{username} was kicked by {kickerUsername}");
-                }
-            }
-            else if (e.message.messageType == Message.MessageTypes.Heartbeat)
-            {
-                if (FrmHolder.hosting)
-                {
-                    FrmHolder.network.BeginWrite(e.client, FrmHolder.network.ComposeMessage(e.client, 0, Message.MessageTypes.Heartbeat, null, null));
-                }
-            }
-            else if (e.message.messageType == Message.MessageTypes.OwnClientId)
-            {
-                FrmHolder.clientId = Convert.ToUInt32(e.message.messageText);
-            }
-            else if (e.message.messageType == Message.MessageTypes.OtherUserLostConnection)
-            {
-                PrintChatMessage($"{e.message.messageText} has lost connection...");
-            }
-            else if (e.message.messageType == Message.MessageTypes.MadeAdmin)
-            {
-                PrintChatMessage($"You have been made an Admin by {e.message.messageText}");
-            }
-            else if (e.message.messageType == Message.MessageTypes.OtherUserMadeAdmin)
-            {
-                string[] parts = e.message.messageText.Split(' ', 2);
-                string username = parts[0];
-                string setterUsername = parts[1];
-                PrintChatMessage($"{username} has been made an Admin by {setterUsername}");
-            }
-            else if (e.message.messageType == Message.MessageTypes.RemovedAdmin)
-            {
-                PrintChatMessage($"You have been removed from Admin by {e.message.messageText}");
-            }
-            else if (e.message.messageType == Message.MessageTypes.OtherUserRemovedAdmin)
-            {
-                string[] parts = e.message.messageText.Split(' ', 2);
-                string username = parts[0];
-                string setterUsername = parts[1];
-                PrintChatMessage($"{username} has been removed from Admin by {setterUsername}");
-            }
-            else if (e.message.messageType == Message.MessageTypes.SendMessageQueue)
-            {
-                e.client.sendingMessageQueue = true;
-                e.client.receivingMessageQueue = false;
-                FrmHolder.network.SendFirstMessageInMessageQueue(e.client);
-            }
-            else if (e.message.messageType == Message.MessageTypes.ConnectionSetupComplete)
-            {
-                e.client.connectionSetupComplete = true;
-            }
-            else if (e.message.messageType == Message.MessageTypes.ClientId)
-            {
-                if (string.IsNullOrWhiteSpace(e.message.messageText) || e.message.messageText == "0")
-                {
-                    e.client.clientId = FrmHolder.network.nextAssignableClientId;
-                    FrmHolder.network.nextAssignableClientId++;
-                    FrmHolder.network.BeginWrite(e.client, FrmHolder.network.ComposeMessage(e.client, 0, Message.MessageTypes.OwnClientId, e.client.clientId.ToString(), null));
-                    e.client.receivedClientId = true;
-                    return;
-                }
-                uint clientId;
-                bool converted = uint.TryParse(e.message.messageText, out clientId);
-                if (converted)
-                {
-                    for (int i = 0; i < FrmHolder.network.connectedClients.Count(); i++)
-                    {
-                        if (clientId == FrmHolder.network.connectedClients[i].clientId)
-                        {
-                            e.client.sessionFirstConnection = false;
-                            e.client = FrmHolder.network.MergeClient(e.client, FrmHolder.network.connectedClients[i]);
-                            break;
-                        }
-                    }
-                    e.client.clientId = clientId;
-                    e.client.receivedClientId = true;
-                }
-            }
-            else if (e.message.messageType == Message.MessageTypes.RequestVersionNumber)
-            {
-                FrmHolder.network.BeginWrite(e.client, FrmHolder.network.ComposeMessage(e.client, 0, Message.MessageTypes.ClientVersionNumber, VersionNumber.applicationVersion, null));
-            }
-            else if (e.message.messageType == Message.MessageTypes.ClientVersionNumber)
-            {
-                e.client.applicationVersionNumber = (e.message.messageText);
-                e.client.receivedApplicationVersionNumber = true;
-                char versionDifference = VersionNumber.CheckVersionCompatibility(VersionNumber.minimumSupportedClientVersion, VersionNumber.maximumSupportedClientVersion, e.client.applicationVersionNumber, VersionNumber.allowClientPreRelease);
-                FrmHolder.network.BeginWrite(e.client, FrmHolder.network.ComposeMessage(e.client, 0, Message.MessageTypes.ServersMinimumSupportedClientVersionNumber, VersionNumber.minimumSupportedClientVersion, null));
-                FrmHolder.network.BeginWrite(e.client, FrmHolder.network.ComposeMessage(e.client, 0, Message.MessageTypes.ServersMaximumSupportedClientVersionNumber, VersionNumber.maximumSupportedClientVersion, null));
-                FrmHolder.network.BeginWrite(e.client, FrmHolder.network.ComposeMessage(e.client, 0, Message.MessageTypes.ServersPreReleaseSupport, VersionNumber.allowClientPreRelease.ToString(), null));
-                FrmHolder.network.BeginWrite(e.client, FrmHolder.network.ComposeMessage(e.client, 0, Message.MessageTypes.ServerVersionNumberCompatibility, versionDifference.ToString(), null));
-                if (versionDifference == '<' || versionDifference == '>')
-                {
-                    FrmHolder.network.connectedClients.Remove(e.client);
-                    if (e.client.sslStream != null)
-                    {
-                        e.client.sslStream.Close();
-                    }
-                }
-            }
-            else if (e.message.messageType == Message.MessageTypes.RequestUsername)
-            {
-                FrmHolder.network.BeginWrite(e.client, FrmHolder.network.ComposeMessage(e.client, 0, Message.MessageTypes.ClientUsername, FrmHolder.username, null));
-            }
-            else if (e.message.messageType == Message.MessageTypes.ClientUsername)
-            {
-                string requestedUsername = e.message.messageText;
-                bool usernameAlreadyInUse = false;
-                for (int i = 0; i < FrmHolder.network.connectedClients.Count(); i++)
-                {
-                    if (string.Equals(FrmHolder.network.connectedClients[i].username, requestedUsername, StringComparison.OrdinalIgnoreCase))
-                    {
-                        if (e.client.clientId != FrmHolder.network.connectedClients[i].clientId)
-                        {
-                            usernameAlreadyInUse = true;
-                            FrmHolder.network.BeginWrite(e.client, FrmHolder.network.ComposeMessage(e.client, 0, Message.MessageTypes.UsernameInUse, null, null));
-                            break;
-                        }
-                    }
-                }
-                if (usernameAlreadyInUse == false)
-                {
-                    e.client.username = requestedUsername;
-                    e.client.receivedUsername = true;
-                }
-            }
-            else if (e.message.messageType == Message.MessageTypes.RequestClientId)
-            {
-                string clientId = FrmHolder.clientId == 0 ? null : FrmHolder.clientId.ToString();
-                FrmHolder.network.BeginWrite(e.client, FrmHolder.network.ComposeMessage(e.client, 0, Message.MessageTypes.ClientId, clientId, null));
-            }
-            else if (e.message.messageType == Message.MessageTypes.ServersMinimumSupportedClientVersionNumber)
-            {
-                string serverMinimumSupportedClientApplicationVersionNumber = e.message.messageText;
-                e.client.serverMinimumSupportedClientApplicationVersionNumber = serverMinimumSupportedClientApplicationVersionNumber;
-            }
-            else if (e.message.messageType == Message.MessageTypes.ServersMaximumSupportedClientVersionNumber)
-            {
-                string serverMaximumSupportedClientApplicationVersionNumber = e.message.messageText;
-                e.client.serverMaximumSupportedClientApplicationVersionNumber = serverMaximumSupportedClientApplicationVersionNumber;
-            }
-            else if (e.message.messageType == Message.MessageTypes.ServersPreReleaseSupport)
-            {
-                bool serverSupportsClientPreReleaseVersionNumbers = e.message.messageText == "0" ? true : false;
-                e.client.serverSupportsClientPreReleaseAppplicationVersionNumber = serverSupportsClientPreReleaseVersionNumbers;
-            }
-            else if (e.message.messageType == Message.MessageTypes.ServerVersionNumberCompatibility)
-            {
-                char clientApplicationNumberServerCompatibility = '=';
-                bool converted = Char.TryParse(e.message.messageText, out clientApplicationNumberServerCompatibility);
-                if (converted)
-                {
-                    e.client.clientToServerVersionNumberCompatibility = clientApplicationNumberServerCompatibility;
-                    bool unsupportedVersion = false;
-                    if (e.client.clientToServerVersionNumberCompatibility == '<')
-                    {
-                        unsupportedVersion = true;
-                    }
-                    else if (e.client.clientToServerVersionNumberCompatibility == '>')
-                    {
-                        unsupportedVersion = true;
-                    }
-                    if (unsupportedVersion)
-                    {
-                        string difference = clientApplicationNumberServerCompatibility == '<' ? "an older" : "a newer";
-                        MessageBox.Show($"You are running {difference} version ({VersionNumber.applicationVersion}) than that which is supported by the server ({e.client.serverMaximumSupportedClientApplicationVersionNumber}).", "Information", MessageBoxButtons.OK, MessageBoxIcon.Information);
-                        BeginDisconnect();
-                    }
-                }
-                else
-                {
-                    MessageBox.Show($"Unable to determine whether the client is on a version supported by the server.", "Warning", MessageBoxButtons.OK, MessageBoxIcon.Warning);
-                    BeginDisconnect();
-                }
-            }
-            else if (e.message.messageType == Message.MessageTypes.ServerVersionNumber)
-            {
-                string serverApplicationVersionNumber = e.message.messageText;
-                e.client.serverApplicationVersionNumber = serverApplicationVersionNumber;
-            }
-            else if (e.message.messageType == Message.MessageTypes.FinishedSendingMessageQueue)
-            {
-                e.client.receivingMessageQueue = false;
-            }
+            FrmHolder.processing.HeartbeatTimeoutEvent += OnHeartbeatTimeoutFailure;
+            FrmHolder.processing.PrintChatMessageEvent += OnPrintChatMessage;
+            FrmHolder.processing.ClearClientListEvent += OnClearClientList;
+            FrmHolder.processing.AddClientToClientListEvent += OnAddClientToClientList;
+            FrmHolder.processing.ShowMessageBoxEvent += OnShowMessagBox;
+            FrmHolder.processing.OpenMainMenuEvent += OnOpenMainMenu;
         }
 
         private void ClearClientList()
@@ -402,7 +60,7 @@ namespace Chat
             {
                 if (FrmHolder.hosting)
                 {
-                    this.ParentForm.Text = $"{FrmHolder.applicationWindowText} - {FrmHolder.username} hosting on {FrmHolder.network.publicIp}";
+                    this.ParentForm.Text = $"{FrmHolder.applicationWindowText} - {FrmHolder.username} hosting on {FrmHolder.processing.publicIp}";
                 }
                 else
                 {
@@ -415,232 +73,9 @@ namespace Chat
             }
         }
 
-        private bool ProcessCommand(string message)
-        {
-            if (message[0] == '/')
-            {
-                if (FrmHolder.hosting == false)
-                {
-                    PrintChatMessage("You must be an admin to execute commands"); //TODO: Allow non-admin commands (e.g. /help)
-                    return true;
-                }
-                string command = message.Substring(1, message.Length - 1);
-                string[] commandParts = command.Split(' ');
-                if (commandParts[0] == "help" || commandParts[0] == "commands")
-                {
-                    RunCommandHelp(commandParts);
-                }
-                else if (commandParts[0] == "kick") //TODO: Admin only
-                {
-                    return RunCommandKick(commandParts);
-                }
-                else if (commandParts[0] == "admin" && FrmHolder.hosting)
-                {
-                    return RunCommandAdmin(commandParts);
-                }
-                return true;
-            }
-            return false; //true if commmand
-        }
-
-        private void NextStepInConnectionSetupAsServer(Client client)
-        {
-            if (FrmHolder.hosting == false || client.connectionSetupComplete)
-            {
-                return;
-            }
-            if (client.receivedApplicationVersionNumber == false)
-            {
-                if (client.requestedApplicationVersionNumber == false)
-                {
-                    FrmHolder.network.BeginWrite(client, FrmHolder.network.ComposeMessage(client, 0, Message.MessageTypes.RequestVersionNumber, null, null));
-                    client.requestedApplicationVersionNumber = true;
-                }
-                return;
-            }
-            if (client.receivedClientId == false)
-            {
-                if (client.requestedClientId == false)
-                {
-                    FrmHolder.network.BeginWrite(client, FrmHolder.network.ComposeMessage(client, 0, Message.MessageTypes.RequestClientId, null, null));
-                    client.requestedClientId = true;
-                }
-                return;
-            }
-            if (client.receivedUsername == false)
-            {
-                if (client.requestedUsername == false)
-                {
-                    FrmHolder.network.BeginWrite(client, FrmHolder.network.ComposeMessage(client, 0, Message.MessageTypes.RequestUsername, null, null));
-                    client.requestedUsername = true;
-                }
-                return;
-            }
-            if (client.sessionFirstConnection)
-            {
-                List<Client> ignoredClients = new List<Client>();
-                ignoredClients.Add(client);
-                PrintChatMessage($"{client.username} connected");
-                FrmHolder.network.SendToAll(ignoredClients, Message.MessageTypes.UserConnected, client.username, null);
-                FrmHolder.network.UpdateClientLists();
-            }
-            client.connectionSetupComplete = true;
-            FrmHolder.network.BeginWrite(client, FrmHolder.network.ComposeMessage(client, 0, Message.MessageTypes.ConnectionSetupComplete, null, null));
-            FrmHolder.network.BeginWrite(client, FrmHolder.network.ComposeMessage(client, 0, Message.MessageTypes.SendMessageQueue, null, null));
-            client.receivingMessageQueue = true;
-        }
-
         private void PrintChatMessage(string chatMessage)
         {
             xlbxChat.Items.Add(chatMessage);
-        }
-
-        private void RunCommandHelp(string[] commandParts)
-        {
-            if (commandParts.Length == 1)
-            {
-                PrintChatMessage($"Available commands are 'kick' and 'admin'. Type '/help [command]' for an explanation of the command.");
-            }
-            else if (commandParts[1] == "kick")
-            {
-                PrintChatMessage($"Explanation: Kicks a user. Format: {kickFormat}");
-            }
-            else if (commandParts[1] == "admin")
-            {
-                PrintChatMessage($"Explanation: Adds/removes a user from Admin. Format: {adminFormat}");
-            }
-            else
-            {
-                PrintChatMessage($"No command {commandParts[1]} exists.");
-            }
-        }
-
-        private bool RunCommandKick(string[] commandParts)
-        {
-            if (commandParts[1] == null)
-            {
-                PrintChatMessage($"The format is: {kickFormat}");
-                return true;
-            }
-            string[] username = { commandParts[1] };
-            List<Client> clients = FrmHolder.network.ClientSearch(username, null);
-            if (clients.Count == 0)
-            {
-                PrintChatMessage($"No user with the username {username[0]} exists");
-                return true;
-            }
-            string reason = "";
-            for (int i = 2; i < commandParts.Length; i++)
-            {
-                reason += commandParts[i] + ' ';
-            }
-            reason = reason.Trim();
-            if (!string.IsNullOrWhiteSpace(reason))
-            {
-                PrintChatMessage($"You kicked {username[0]} with reason: {reason}");
-            }
-            else
-            {
-                PrintChatMessage($"You kicked {username[0]}");
-            }
-            List<Client> ignoredClients = new List<Client>();
-            ignoredClients.Add(clients[0]);
-            FrmHolder.network.BeginWrite(clients[0], FrmHolder.network.ComposeMessage(clients[0], 0, Message.MessageTypes.Kicked, $"{FrmHolder.username} {reason}", null)); // Kick client
-            FrmHolder.network.SendToAll(ignoredClients, Message.MessageTypes.OtherUserKicked, $"{username[0]} {FrmHolder.username} {reason}", null);
-            return false;
-        }
-
-        private bool RunCommandAdmin(string[] commandParts)
-        {
-            if (commandParts.Length < 2 || commandParts[1] == null)
-            {
-                PrintChatMessage($"The format is: {adminFormat}");
-                return true;
-            }
-            string[] username = { commandParts[1] };
-            List<Client> clients = FrmHolder.network.ClientSearch(username, null);
-            if (clients.Count == 0)
-            {
-                PrintChatMessage($"No user with the username {username[0]} exists");
-                return true;
-            }
-            bool setAsAdmin = false;
-            if (commandParts.Length > 2 && commandParts[2] != null)
-            {
-                if (string.Equals(commandParts[2], "True", StringComparison.OrdinalIgnoreCase))
-                {
-                    if (clients[0].admin)
-                    {
-                        PrintChatMessage($"This user is already an Admin");
-                        return true;
-                    }
-                    setAsAdmin = true;
-                }
-                else if (string.Equals(commandParts[2], "False", StringComparison.OrdinalIgnoreCase))
-                {
-                    if (clients[0].admin == false)
-                    {
-                        PrintChatMessage($"This user is already not an Admin");
-                        return true;
-                    }
-                    setAsAdmin = false;
-                }
-                else
-                {
-                    PrintChatMessage($"The format is: {adminFormat}");
-                    return true;
-                }
-            }
-            else
-            {
-                if (FrmHolder.hosting)
-                {
-                    if (FrmHolder.network.ClientSearch(username, null)[0].admin)
-                    {
-                        setAsAdmin = false;
-                    }
-                    else
-                    {
-                        setAsAdmin = true;
-                    }
-                }
-            }
-            if (FrmHolder.hosting)
-            {
-                clients[0].admin = setAsAdmin;
-                FrmHolder.network.SetAdmin(clients[0], FrmHolder.username, setAsAdmin, null);
-            }
-            return false;
-        }
-
-        private bool BeginDisconnect()
-        {
-            bool serverClose = false;
-            if (FrmHolder.hosting)
-            {
-                DialogResult dialogResult = MessageBox.Show("This will terminate the server. Are you sure?", "Warning", MessageBoxButtons.OKCancel, MessageBoxIcon.Warning);
-                if (dialogResult == DialogResult.OK)
-                {
-                    serverClose = true;
-                    if (FrmHolder.network.serverThread != null && FrmHolder.network.serverThread.IsAlive)
-                    {
-                        FrmHolder.network.serverCancellationTokenSource.Cancel();
-                    }
-                }
-                else if (dialogResult == DialogResult.Cancel)
-                {
-                    return false;
-                }
-            }
-            if (FrmHolder.network.clientThread != null && FrmHolder.network.clientThread.IsAlive)
-            {
-                FrmHolder.network.clientCancellationTokenSource.Cancel();
-            }
-            if (serverClose || FrmHolder.hosting == false)
-            {
-                OpenMainMenu();
-            }
-            return true;
         }
 
 
@@ -669,27 +104,7 @@ namespace Chat
         {
             if (xlbxChat.InvokeRequired)
             {
-                xlbxChat.BeginInvoke(new HeartbeatDelegate(FrmHolder.network.HeartbeatTimeoutFailure), client);
-            }
-        }
-
-        private void OnMessageReceived(object sender, MessageReceivedEventArgs e)
-        {
-            if (xlbxChat.InvokeRequired)
-            {
-                xlbxChat.BeginInvoke(new MessageDelegate(ProcessMessage), this, e);
-                if (e.client.connectionSetupComplete == false)
-                {
-                    xlbxChat.BeginInvoke(new NextConnectionSetupStepDelegate(NextStepInConnectionSetupAsServer), e.client);
-                }
-            }
-            else
-            {
-                ProcessMessage(this, e);
-                if (e.client.connectionSetupComplete == false)
-                {
-                    NextStepInConnectionSetupAsServer(e.client);
-                }
+                xlbxChat.BeginInvoke(new HeartbeatDelegate(FrmHolder.processing.HeartbeatTimeoutFailure), client);
             }
         }
 
@@ -729,18 +144,6 @@ namespace Chat
             }
         }
 
-        private void OnNextConnectionSetupStep(object sender, Client client)
-        {
-            if (xlbxChat.InvokeRequired)
-            {
-                xlbxChat.BeginInvoke(new NextConnectionSetupStepDelegate(NextStepInConnectionSetupAsServer), client);
-            }
-            else
-            {
-                NextStepInConnectionSetupAsServer(client);
-            }
-        }
-
         private void OnShowMessagBox(object sender, ShowMessageBoxEventArgs showMessageBoxEventArgs)
         {
             if (this.InvokeRequired)
@@ -753,11 +156,23 @@ namespace Chat
             }
         }
 
+        private void OnOpenMainMenu(object sender, EventArgs e)
+        {
+            if (this.InvokeRequired)
+            {
+                this.BeginInvoke(new OpenMainMenuDelegate(OpenMainMenu));
+            }
+            else
+            {
+                OpenMainMenu();
+            }
+        }
+
         private void OnClosing(object sender, FormClosingEventArgs e)
         {
             if (askToClose) // Prevents closing when returning to main menu
             {
-                if (BeginDisconnect() == false)
+                if (FrmHolder.processing.BeginDisconnect() == false)
                 {
                     e.Cancel = true;
                 }
@@ -778,14 +193,14 @@ namespace Chat
                     {
                         PrintChatMessage($"{FrmHolder.username}: {message}");
                     }
-                    if (ProcessCommand(message) == false)
+                    if (FrmHolder.processing.ProcessCommand(message) == false)
                     {
-                        if (FrmHolder.network.connectedClients.Count > 0)
+                        if (FrmHolder.processing.connectedClients.Count > 0)
                         {
                             message = message.Trim();
-                            for (int i = 0; i < FrmHolder.network.connectedClients.Count; i++)
+                            for (int i = 0; i < FrmHolder.processing.connectedClients.Count; i++)
                             {
-                                FrmHolder.network.BeginWrite(FrmHolder.network.connectedClients[i], FrmHolder.network.ComposeMessage(FrmHolder.network.connectedClients[i], 0, Message.MessageTypes.ChatMessage, $"{FrmHolder.username} {message}", null));
+                                FrmHolder.processing.BeginWrite(FrmHolder.processing.connectedClients[i], FrmHolder.processing.ComposeMessage(FrmHolder.processing.connectedClients[i], 0, Message.MessageTypes.ChatMessage, $"{FrmHolder.username} {message}", null));
                             }
                         }
                     }
@@ -815,7 +230,7 @@ namespace Chat
 
         private void xbtnDisconnect_Click(object sender, EventArgs e)
         {
-            BeginDisconnect();
+            FrmHolder.processing.BeginDisconnect();
         }
         #endregion
     }
